// utils/enums.ts
export enum CompanyMemberRole {
  ADMIN = "ADMIN",
  MEMBER = "MEMBER",
  ACCOUNTING = "ACCOUNTING",
  ACCOUNTING_HEAD = "ACCOUNTING_HEAD",
  MERCHANT = "MERCHANT",
}

export enum RouteAction {
  ALLOW = "ALLOW",
  REDIRECT_LOGIN = "REDIRECT_LOGIN",
  REDIRECT_DASHBOARD = "REDIRECT_DASHBOARD",
  FORBIDDEN = "FORBIDDEN",
}